Redesign DOMList and review its implementations.
DOMList is now no more than an interface for things that have a
`length` and index properties. We add a Seq view of all DOMLists
in its companion object. We make it read-only and covariant, as
its implementations are actually not mutable in the DOM standard.

We review all its subclasses and subtraits:

* Make them classes iff they are exposed and can be used with
  `instanceof`. Their constructors are always private. Otherwise,
  make them traits.
* Explicitly declare `item()` iff it is supported in the given
  class/trait (it used to be in `DOMList`, but not all
  implementations support `item()`).

We make `NodeList` parameterized and covariant in the type of node.
We deprecate `NodeListOf` in favor of `NodeList`.